issue-183 Connect recipe page to database
