+++ prisma.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig } from "prisma/config";
+import { defineConfig } from 'prisma/config';
 import 'dotenv/config';
 
 export default defineConfig({
   migrations: {
-    seed: "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts",
+    seed: 'ts-node --compiler-options {"module":"CommonJS"} prisma/seed.ts',
   },
 });
